==> 02_draft_notebook.ipynb <==
# Draft notebook:
In this notebook we carry some more pre-processing on the data generate through 03_google_trends_notebook and 04_bitcoin_data_notebook. And then we load the data from bitcointalk.org forum (private source not share on this git), and carry Sentiment Analysis on it.

# 0. Preprocessing

import pandas as pd
import numpy as np
import nltk
import matplotlib.pyplot as plt
import seaborn as sns
pd.set_option('display.max_rows', 15)

### 0.1 Prices and returns data

poloniex_data = pd.read_csv('Data/poloniex_data.csv', index_col = 'date')

poloniex_data

### 0.2 Google trends data: buy over buy + sell ratio (buy percentage)

google_trends_data = pd.read_csv('Data/google_trends.csv', index_col = 'date')

google_trends_data.drop("Unnamed: 0", axis =1)

### 0.3 Bitcointalk forum dataset

forum_data = pd.read_csv('Data/NLP_Cryptocurrency_Dataset.csv', header = 1)

forum_data = forum_data.drop('Unnamed: 0', axis = 1)

forum_data.datetime = pd.to_datetime(forum_data.datetime)

forum_data_prep = forum_data.set_index('datetime').copy()

forum_data_prep = forum_data_prep["may-2017":"may-2018"].dropna()

forum_data_prep = forum_data_prep.reset_index()

forum_data_prep

# 1. Sentiment Analysis

https://www.learndatasci.com/tutorials/sentiment-analysis-reddit-headlines-pythons-nltk/

from nltk.sentiment.vader import SentimentIntensityAnalyzer as SIA

sia = SIA()
results = []

for line in forum_data_prep.post:  #[:1000]
    pol_score = sia.polarity_scores(line)
    pol_score['post'] = line
    results.append(pol_score)

results[:3]

forum_sentiment = pd.DataFrame.from_records(results)

forum_sentiment

forum_sentiment.to_csv('Data/forum_sentiment.csv')

forum_sentiment

forum_sentiment.columns = ["compound","neg","neu","pos","comment"]

forum_NLP = pd.concat([forum_data_prep, forum_sentiment], axis = 1)

forum_NLP = forum_NLP.drop("comment", axis = 1)

forum_NLP

forum_NLP.to_csv('Data/forum_NLP.csv')
