==> 03_google_trends_notebook.ipynb <==
# Google Trends Momentum

In this notebook we download data from Google trends and carry out some of the pre-processing related to it. What we look at is the ratio of "buy bitcoin" versus the sum of "buy bitcoin" & "sell bitcoin" to get a feeling about how the trend is evolving over time.

import pandas as pd
from pytrends.request import TrendReq

timeframes2017 = list()
timeframesmix = list()
timeframes2018 = list()
for i in range(5,11):
    timeframes2017.append('2017-'+str(i)+'-1 2017-'+str(i+2)+'-1')
timeframesmix.append('2017-'+'11'+'-1 2018-'+'1'+'-1')
timeframesmix.append('2017-'+'12'+'-1 2018-'+'2'+'-1')
for i in range(1,5):
    timeframes2018.append('2018-'+str(i)+'-1 2018-'+str(i+2)+'-1')

timeframes = timeframes2017+timeframesmix+timeframes2018

pytrend = TrendReq(hl='en-US', tz=360)
list_google = list()
for timeframe in timeframes:
    pytrend.build_payload(kw_list=['Buy Bitcoin', 'Sell Bitcoin'], 
                      timeframe=timeframe)
    print(timeframe)
    list_google.append(pytrend.interest_over_time())

df = pd.DataFrame(list_google[0])

for i in range(1,11):
    df = df.append(list_google[i])

df['googletrends_buy_sell'] = df["Buy Bitcoin"]/(df["Sell Bitcoin"]+df["Buy Bitcoin"])
df = df[['googletrends_buy_sell']] 

df = df.groupby('date').mean().reset_index()
df.to_csv("Data/google_trends.csv")
