==> 04_bitcoin_data_notebook.ipynb <==
# Bitcoin prices notebook
In this notebook we download bitcoin prices timeseries from Poloniex and carry out some of the pre-processing.

## Step 0. Import usual libraries

import pandas as pd
import numpy as np
import poloniex
import datetime
#import ffn

pd.set_option('display.max_rows', 15)

## Step 1. Importing from Poloniex
Poloniex is a cryptocurrency exchange found on https://poloniex.com/exchange, to download the data we use a package made by a poloniex community member.

Importing from API

* To import poloniex we need to install the package in the console using:

`pip install poloniex`
* The help function gives a list of functions included in the package and some descriptions

`help(poloniex.poloniex)`
* We are using the public data so no keys are needed

The package has many functionalities but the information we want is public, therefore we do not need access tokens or secret keys. The currency we are looking for is BitCoin relative to the USD, therefore looking at the documentation we know that we need to search using the term `'USDT_BTC'`, with an additional parameter of seconds between measurements. We chose to take the 24 hours frequency.

polo = poloniex.Poloniex()
#btc300 = pd.DataFrame(polo.returnChartData("USDT_BTC", 300))
btc86400 = pd.DataFrame(polo.returnChartData("USDT_BTC", 86400))

Quick look at the data:

btc86400

## Step 2. Pre-processing

Transform the unix timestamp to datetime:

btc86400.date = pd.to_datetime(btc86400.date, unit = 's')

Setting the date as index:

btc86400.set_index('date', inplace = True)

Taking the period 2017-05-01 to 2018-05-02 (one more day for lagged purposes):

btc_data = btc86400["2017-04-16":"2018-05-02"]
btc_data

We decide to take only close and volume since the other columns are extremly correlated to these and may not provide more useful information for our purpose.

returns_data = btc_data[["close"]]

Now let's look at the returns (indeed, the close price itself is not particularly interesting):

returns_data = returns_data.pct_change()

returns_data

We are going to add the volatility of the last 14 days:

volatility_data = returns_data.rolling(14).std()
volatility_data = volatility_data.rename(columns = {'close': 'volatility_14'})
volatility_data

returns_data = returns_data.shift(-1)
returns_data = returns_data.dropna()

Let's rename the columns for later purposes:

returns_data.columns = ["return_day+1"]

returns_data

Now we put back together the returns at day+1 with the close and volume information from earlier.
We also dropna() for deleting the day 2018-05-02.

data = pd.concat((returns_data, btc_data[["close","volume"]]), axis = 1)
data = pd.concat((data, volatility_data), axis = 1)

data = data.dropna()
data = data["may-2017":"may-2018"]

Creating a csv file for later purposes and checking that reading it works.

data.to_csv("Data/poloniex_data.csv")

data = pd.read_csv("poloniex_data.csv", index_col = 'date')

data
