==> 05_twitter_data_notebook.ipynb <==
# TWITTER WEBSCRAPPING

Here we are going to webscrap the twitter accounts of the top-100 influencers in the cryptocurrency world. The first step is webscrapping the top100 accounts and then all the tweets they perfromed.

#### Data Sources:

Cryptocurrencies influencer ranking: "https://cryptoweekly.co/100/"

The following code has been adapted from:
https://codeburst.io/a-twitter-analysis-of-the-100-most-influential-people-in-crypto-bb95b2608925 and https://dev.to/rodolfoferro/sentiment-analysis-on-trumpss-tweets-using-python-

# First we load the general packages:
import sys
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

# Here we are going to write a function to webscarpe the names of the influencers and their twitter handles:

from urllib.request import urlopen
from bs4 import BeautifulSoup as soup

def getTwitterHandles():
	# Fill in with url of page which is to be scraped
	url = "https://cryptoweekly.co/100/"

	# Retreives and parses page html
	client = urlopen(url)
	pageHtml = client.read()
	pageSoup = soup(pageHtml, "html.parser")

	# Adds all Twitter handles to twitterHandles list
	profiles = pageSoup.findAll("div", {"class":"testimonial-wrapper"})
	twitterHandles = []
	for person in profiles:
		twitterHandles.append(person.findAll("div",{"class":"author"}))
	for i in range(len(twitterHandles)):
		twitterHandles[i]=twitterHandles[i][0].findAll("a")[0].text[1:]

	client.close()
	return twitterHandles


# Here the Twitter API tweepy is used to webscrape the tweets from the influencers:

import tweepy
import csv
import sys

# Twitter API credentials (expired, don't even try it)
consumer_key = ""
consumer_secret = ""
access_key = ""
access_secret = ""


def get_all_tweets(screen_name):
	print("Getting tweets from @" + str(screen_name))

	#Twitter only allows access to a users most recent 3240 tweets with this method
	
	#authorize twitter, initialize tweepy
	auth = tweepy.OAuthHandler(consumer_key, consumer_secret)
	auth.set_access_token(access_key, access_secret)
	api = tweepy.API(auth)
	
	#initialize a list to hold all the tweepy Tweets
	alltweets = []	
	
	#make initial request for most recent tweets (200 is the maximum allowed count)
	new_tweets = api.user_timeline(screen_name = screen_name,count=200)
	
	#save most recent tweets
	alltweets.extend(new_tweets)
	
	#save the id of the oldest tweet less one
	oldest = alltweets[-1].id - 1
	
	#keep grabbing tweets until there are no tweets left to grab
	while len(new_tweets) > 0:
		print ("Getting tweets before %s" % (oldest))
		
		#all subsiquent requests use the max_id param to prevent duplicates
		new_tweets = api.user_timeline(screen_name = screen_name,count=200,max_id=oldest)
		
		#save most recent tweets
		alltweets.extend(new_tweets)
		
		#update the id of the oldest tweet less one
		oldest = alltweets[-1].id - 1
		
		print ("...%s tweets downloaded so far" % (len(alltweets)))

	#transform the tweepy tweets into a 2D array that will populate the csv	
	outtweets = [[tweet.id_str.encode('utf-8'), tweet.created_at.strftime('%m/%d/%Y'), tweet.text.encode('utf-8')] for tweet in alltweets]
	
	#write the csv	
	with open('C:/Users/Eric/Python Projects/Bitcoin-NLP-Strategy/Data/Tweets/%s_tweets.csv' % screen_name, 'w') as f:
		writer = csv.writer(f)
		writer.writerow(["id","created_at","text"])
		writer.writerows(outtweets)
	
	pass


if __name__ == '__main__':
	handles = getTwitterHandles()
    # The Peter Todd twitter handle is updated to his new account, as long as the old one was listed in the website:
	for i in range(len(handles)):
		if handles[i] == 'petertoddbtc':
			handles[i] = 'peterktodd'
	for handle in handles:
		get_all_tweets(str(handle))

# The previous function generated a .csv file for every account, here we create a unified dataset:
# Create a unified dataset:
mt = pd.DataFrame(columns = ["Name", "id", "created_at", "text"])

for handle in handles:
    df = pd.read_csv("Data/Tweets/"+str(handle)+"_tweets.csv")
    df['Name'] = handle
    df = df[["Name", "id", "created_at", "text"]]
    mt = mt.append(df)

pd.DataFrame.to_csv(mt, "Data/twitter_ddbb.csv")

## TWITTER SENTIMENT ANALYSIS

from textblob import TextBlob
import re

def clean_tweet(tweet):
    return ' '.join(re.sub("(@[A-Za-z0-9]+)|([^0-9A-Za-z \t])|(\w+:\/\/\S+)", " ", tweet).split())

def analize_sentiment(tweet):
    analysis = TextBlob(clean_tweet(tweet))
    return analysis.sentiment.polarity    
    
# We create a column with the result of the analysis:
mt['SA'] = np.array([ analize_sentiment(tweet) for tweet in mt['text']])

# We display the updated dataframe with the new column:
#display(mt.head(10))

pd.DataFrame.to_csv(mt, "Data/twitter_ddbb_SA.csv")

Let's reaload the data again, for further preprocessing without need of running the scrape and NLP processes every time:

mt = pd.read_csv("Data/twitter_ddbb_SA.csv", index_col=0)

# Selecting only the tweets in the desired time frame:
mt['created_at'] = pd.to_datetime(mt['created_at'])
mask = (mt['created_at'] >= '05/01/2017') & (mt['created_at'] <= '05/01/2018')
mt.loc[mask]
twind_mt = mt.loc[mask]

# Group by DAY and aggreagate the data:
agg_mt = twind_mt.groupby('created_at').agg({'text':'count', 'Name': 'nunique', 'SA': 'mean'})
agg_mt = agg_mt.rename(columns={'text': 'Tweets (#)', 'Name': 'Active Influencers (#)', 'SA': 'Average SA'})

pd.DataFrame.to_csv(agg_mt, "C:/Users/Eric/Python Projects/Bitcoin-NLP-Strategy/Data/twitter_agg_ddb.csv")
