==> 06_reddit_api_data_notebook.ipynb <==
# Reddit API data fetching
In this notebook we fetch the top threads about Bitcoin in reddit.

import praw
from praw.models import MoreComments
import pandas as pd
reddit = praw.Reddit(client_id='4qQrwmPQAVOwiw', \
                     client_secret='MVf28OrOH-Ak_-XLSlLHWqbfono', \
                     user_agent='API dan', \
                     username='blablabla123983', \
                     password='newguy')
headlines = set()

topics_dict = {
                "date": [], \
                "body":[],\
                "popularity":[]
              }

for submission in reddit.subreddit('bitcoin').top(limit=1500):
#for submission in reddit.subreddit('bitcoin').top('all'):
    if isinstance(submission, MoreComments):
        continue
    topics_dict["date"].append(submission.created)
    topics_dict["body"].append(submission.title)
    topics_dict["popularity"].append(submission.num_comments)

reddit_data = pd.DataFrame
reddit_data = pd.DataFrame.from_dict(topics_dict)
#reddit_data = reddit_data.set_index('date')

reddit_data.date = pd.to_datetime(reddit_data.date, unit = 's')

reddit_data = reddit_data[(reddit_data['date'] > '2017-05-01')]
reddit_data

reddit_data = reddit_data.set_index('date')

reddit_data

reddit_data.to_csv('bitcoin_reddit.csv', index=False, sep=";")
